CLI-1488 CLI-1593
Implement post-commit artefact refresh for changed files

This commit introduces a new function, `run_post_commit_artefact_refresh`, which processes files changed in a commit, updating the DevQL database with their states. It tracks statistics such as files seen, indexed, deleted, and failed during the refresh process. Additionally, the `ManualCommitStrategy` is updated to call this new function after a commit, ensuring that the current state reflects the latest changes. Tests are added to verify the correct indexing of changed files and the removal of deleted files from the database.
Broke down post-commit to submodules.
